requirements.txt
anyio==4.6.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.6
python-dotenv==0.19.0
python-telegram-bot==21.6
pytz==2024.1
requests==2.31.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
tornado==6.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
